Pullup r19183:r19510 from trunk
git-svn-id: svn://anonsvn.mit.edu/krb5/users/coffman/pkinit@19511 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/os/an_to_ln.c
@@ -47,7 +47,7 @@
 #include <regexpr.h>
 #endif	/* !HAVE_REGCOMP && HAVE_REGEXP_H && HAVE_COMPILE */
 
-#define	MAX_FORMAT_BUFFER	1024
+#define	MAX_FORMAT_BUFFER	((size_t)1024)
 #ifndef	min
 #define	min(a,b)	((a>b) ? b : a)
 #endif	/* min */
